requirements.txt
coloredlogs==15.0.1
Flask==3.1.2
pycryptodome==3.20.0
PyJWT==2.10.1
pymongo==4.15.5
Werkzeug==3.1.4
